--- src/main/java/com/nhnacademy/domain/UserBirthEvent.java
@@ -1,18 +1,0 @@
-package com.nhnacademy.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.io.Serializable;
-import java.time.LocalDate;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserBirthEvent implements Serializable {
-    private Long userNo;
-    private LocalDate userBirth;
-}
