+++ src/main/java/com/nhnacademy/domain/enumtype/CouponDiscountType.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.domain;
+package com.nhnacademy.domain.enumtype;
 
 public enum CouponDiscountType {
     PERCENT,     // 백분율
