+++ src/main/java/com/nhnacademy/domain/enumtype/CouponScope.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.domain;
+package com.nhnacademy.domain.enumtype;
 
 public enum CouponScope {
     ALL,
