+++ src/main/java/com/nhnacademy/dto/response/CouponPolicyDetailResponse.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.dto.response;
 
-import com.nhnacademy.domain.CouponDiscountType;
-import com.nhnacademy.domain.CouponScope;
+import com.nhnacademy.domain.enumtype.CouponDiscountType;
+import com.nhnacademy.domain.enumtype.CouponScope;
 import com.nhnacademy.domain.CouponPolicy;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
