+++ src/main/java/com/nhnacademy/listener/BirthdayEventListener.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.listener;
 
-import com.nhnacademy.domain.UserBirthEvent;
+import com.nhnacademy.event.UserBirthEvent;
 import com.nhnacademy.service.CouponService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
