+++ src/main/java/com/nhnacademy/repository/CouponBookRepository.java
@@ -4,11 +4,10 @@ import com.nhnacademy.domain.CouponBook;
 import com.nhnacademy.repository.queryfactory.CouponBookRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 
-public interface CouponBookRepository extends JpaRepository<CouponBook, Long>, QuerydslPredicateExecutor<CouponBook>, CouponBookRepositoryCustom {
+public interface CouponBookRepository extends JpaRepository<CouponBook, Long>, CouponBookRepositoryCustom {
 
     List<CouponBook> findByCouponPolicy_CouponId(Long couponId);
