+++ src/main/java/com/nhnacademy/repository/CouponCategoryRepository.java
@@ -4,13 +4,11 @@ import com.nhnacademy.domain.CouponCategory;
 import com.nhnacademy.repository.queryfactory.CouponCategoryRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 
-public interface CouponCategoryRepository extends JpaRepository<CouponCategory, Long>,
-        QuerydslPredicateExecutor<CouponCategory>,
-        CouponCategoryRepositoryCustom {
+public interface CouponCategoryRepository extends JpaRepository<CouponCategory, Long>, CouponCategoryRepositoryCustom {
+
     List<CouponCategory> findByCouponPolicy_CouponId(Long couponId);
 
     @Query("SELECT cc.categoryId FROM CouponCategory cc WHERE cc.couponPolicy.couponId = :couponId")
