+++ src/main/java/com/nhnacademy/repository/CouponPolicyRepository.java
@@ -3,10 +3,6 @@ package com.nhnacademy.repository;
 import com.nhnacademy.domain.CouponPolicy;
 import com.nhnacademy.repository.queryfactory.CouponPolicyRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface CouponPolicyRepository extends JpaRepository<CouponPolicy, Long>,
-        QuerydslPredicateExecutor<CouponPolicy>,
-        CouponPolicyRepositoryCustom {
-
+public interface CouponPolicyRepository extends JpaRepository<CouponPolicy, Long>, CouponPolicyRepositoryCustom {
 }
