+++ src/main/java/com/nhnacademy/repository/UserCouponRepository.java
@@ -4,14 +4,10 @@ import com.nhnacademy.domain.CouponPolicy;
 import com.nhnacademy.domain.UserCoupon;
 import com.nhnacademy.repository.queryfactory.UserCouponRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-
 import java.util.List;
 import java.util.Optional;
 
-public interface UserCouponRepository extends JpaRepository<UserCoupon, Long>,
-        QuerydslPredicateExecutor<UserCoupon>,
-        UserCouponRepositoryCustom {
+public interface UserCouponRepository extends JpaRepository<UserCoupon, Long>, UserCouponRepositoryCustom {
 
     //추후에 사용 예정
 //    List<UsedCoupon> findByUserNo(Long userNo);
