+++ src/main/resources/application-dev.properties
@@ -1,10 +1,4 @@
-
 spring.datasource.url=jdbc:mysql://10.116.64.14:13306/project_be10_team3_dev?useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-
-# 민감 정보를 api에 명시적으로 적어두는게 아닌 config-server 를 통해 불러와야하는데 일단 개발중이니 주석 처리함
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.username=project_be10_team3
-spring.datasource.password=5T0/S)(BKbgiLhac
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
