+++ src/test/java/com/nhnacademy/BookstoreCouponApiApplicationTests.java
@@ -2,10 +2,8 @@ package com.nhnacademy;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-@ActiveProfiles("test")
 class BookstoreCouponApiApplicationTests {
 
     @Test
