--- src/test/resources/data.sql
@@ -1,1 +1,0 @@
-SELECT 1;
